[ifta] enhancement(core): refactor IFTA page to feature (wave-#1) (#123)
